FreeRTOS configuration changes
tried adding a task for the RTOS and got erros soled that, then noticed an error with the button presses getting a hard fault soled. will uncomment the  code in the task once i get check its working as expected

configTOTAL_HEAP_SIZE this was needed to allow more than one task as previoulsy the heap was too small

configTIMER_TASK_STACK_DEPTH there was an odd error when pressing buttons on the dev board that caues a hard fault increasing this took care of it.