i2c eeprom write/read
adding i2c(twi) read/write, was a bit tricky to get working as i was getting a NACK(negative acknolwdgement) on the second write that is why there are do whiles and delays.

also change a vat to volitile since it is used in an interrupt handler.

also some cleaning up of defined values for move to PCB schematic