change morse code speed
change the speed of morse code, use menu option 'S' to cycle through some changes. and then change it. this only changes the space between letters and for spaces. so each character will remain un changed.

also changed send message task so when you try to queue a message it will only allow you to add more to a queue once the message is complete.